ML_SentimentAnalysis/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
joblib==1.4.2
nltk==3.9.1
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sklearn-pandas==2.2.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.3.0
vaderSentiment==3.3.2
langdetect==1.0.9